# admin/jwtdemo/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Simple JWT authentication example with Swagger UI (Gin + swaggo).",
        "title": "JWT Auth Demo API",
        "contact": {
            "name": "API Support",
            "email": "you@example.com"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Requires Bearer token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Get profile (protected)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Profile"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/main.APIError"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Simple health endpoint",
                "tags": [
                    "public"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Use demo creds to get a JWT: username=admin, password=admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login (demo)",
                "parameters": [
                    {
                        "description": "Credentials",
                        "name": "creds",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.TokenResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/main.APIError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.APIError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "invalid credentials"
                }
            }
        },
        "main.LoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "admin"
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
        "main.Profile": {
            "type": "object",
            "properties": {
                "expires": {
                    "type": "string",
                    "example": "2025-09-26T13:00:00Z"
                },
                "issuer": {
                    "type": "string",
                    "example": "jwtdemo"
                },
                "role": {
                    "type": "string",
                    "example": "demo-admin"
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
        "main.TokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "\u003cjwt\u003e"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter: Bearer {your JWT}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}